Create Python for Finance Brandon
What Is Python?
Python is a high-level, multipurpose programming language that is used in a wide range of
domains and technical fields. On the Python website you find the following executive
summary (cf. https://www.python.org/doc/essays/blurb):
Python is an interpreted, object-oriented, high-level programming language with dynamic semantics. Its highlevel
built in data structures, combined with dynamic typing and dynamic binding, make it very attractive for
Rapid Application Development, as well as for use as a scripting or glue language to connect existing components
together. Python’s simple, easy to learn syntax emphasizes readability and therefore reduces the cost of program
maintenance. Python supports modules and packages, which encourages program modularity and code reuse. The
Python interpreter and the extensive standard library are available in source or binary form without charge for all
major platforms, and can be freely distributed.
This pretty well describes why Python has evolved into one of the major programming
languages as of today. Nowadays, Python is used by the beginner programmer as well as
by the highly skilled expert developer, at schools, in universities, at web companies, in
large corporations and financial institutions, as well as in any scientific field.
Among others, Python is characterized by the following features:
Open source
Python and the majority of supporting libraries and tools available are open source
and generally come with quite flexible and open licenses.
Interpreted
The reference CPython implementation is an interpreter of the language that
translates Python code at runtime to executable byte code.
Multiparadigm
Python supports different programming and implementation paradigms, such as
object orientation and imperative, functional, or procedural programming.
Multipurpose
Python can be used for rapid, interactive code development as well as for building
large applications; it can be used for low-level systems operations as well as for highlevel
analytics tasks.
Cross-platform
Python is available for the most important operating systems, such as Windows,
Linux, and Mac OS; it is used to build desktop as well as web applications; it can be
used on the largest clusters and most powerful servers as well as on such small
devices as the Raspberry Pi (cf. http://www.raspberrypi.org).
Dynamically typed
Types in Python are in general inferred during runtime and not statically declared as
in most compiled languages.
Indentation aware
In contrast to the majority of other programming languages, Python uses indentation
www.it-ebooks.info
for marking code blocks instead of parentheses, brackets, or semicolons.